get coverage information to be compressed 100x more than it was in run log file
Coverage information is now saved differently in the collection
phase. Before, we were saving the (almost) full `Coverage.peek_result` (albeit
modified for compactness) in the run log file. During the nth test method, the
before and after coverage was compacted, filtered, and diffed with the coverage
collected BEFORE any test methods were run.

Now, during the nth test method, the same process occurs except it is now diffed
with the coverage collected in the PREVIOUS  (n-1) test method.

This assumes certain things: that either the test suite is ordered (probably
not), or that the suite is using a random seeded order, so that during playback
of the selected tests (`covet --exec`), we can pick the same ordering (the
seed).

You might ask why we even need to preserve the test ordering at all, since we
have the coverage information of each test method, which presumably would be the
same regardless of test ordering. This isn't true, however, because of
`autoload`. If a library uses `autoload`, a constant's first reference can vary,
which varies the loading of files and thus the coverage information gathered
during test methods.